Adding job removal functions. Adding more buttons to UI. Fixing dbpatch. Fixing handling of timestamps.

+++ lib/SGN/Controller/AJAX/Job.pm
@@ -19,7 +19,11 @@ sub retrieve_jobs_by_user :Path('/ajax/job/jobs_by_user') Args(1) {
     my $c = shift;
     my $sp_person_id = shift;
 
-    #check logged in user and proceed if ok
+    my $logged_user = $c->user() ? $c->user->get_object()->get_sp_person_id() : undef;
+    my $role = $c->user() ? $c->user->get_object()->get_user_type() : undef;
+    if ($sp_person_id ne $logged_user && $role ne "curator") {
+        die "You do not have permission to see these job logs.\n";
+    }
 
     my $bcs_schema = $c->dbic_schema("Bio::Chado::Schema");
     my $people_schema = $c->dbic_schema("CXGN::People::Schema");
@@ -47,17 +51,15 @@ sub retrieve_jobs_by_user :Path('/ajax/job/jobs_by_user') Args(1) {
         my $job = CXGN::Job->new({
             schema => $bcs_schema,
             people_schema => $people_schema,
-            sp_job_id => $job_id,
-            args => {
-                logfile => $c->config->{job_finish_log}
-            }
+            sp_job_id => $job_id
         });
-        my $actions_html = "<span id=\"$job_id\" style=\"display: none;\"></span><button id=\"dismiss_job_$job_id\" class=\"btn btn-small btn-danger\">Dismiss</button>";
+        my $actions_html = "<span id=\"$job_id\" style=\"display: none;\"></span><button id=\"dismiss_job_$job_id\" onclick=\"jsMod['job'].dismiss_job($job_id);\" class=\"btn btn-small btn-danger\">Dismiss</button>";
         my $status = $job->check_status();
-        if ($status eq "finished" && $job->retrieve_argument('type') =~ /analysis/) {
-            $actions_html .= "<button id=\"save_job_$job_id\" class=\"btn btn-small btn-success\">";
-        } elsif ($status eq "submitted") {
-            $actions_html .= "<button id=\"cancel_job_$job_id\" class=\"btn btn-small btn-danger\">Cancel</button>"
+        # if ($status eq "finished" && $job->retrieve_argument('type') =~ /analysis/) {
+        #     $actions_html .= "<button id=\"save_job_$job_id\" class=\"btn btn-small btn-success\">Save Results</button>";
+        # } 
+        if ($status eq "submitted") {
+            $actions_html .= "<button id=\"cancel_job_$job_id\" onclick=\"jsMod['job'].cancel_job($job_id)\" class=\"btn btn-small btn-danger\">Cancel</button>"
         }
         my $row = {
             id => $job_id,
@@ -80,6 +82,9 @@ sub delete :Path('/ajax/job/delete') Args(1) {
     my $c = shift;
     my $sp_job_id = shift;
 
+    my $logged_user = $c->user() ? $c->user->get_object()->get_sp_person_id() : undef;
+    my $role = $c->user() ? $c->user->get_object()->get_user_type() : undef;
+
     my $bcs_schema = $c->dbic_schema("Bio::Chado::Schema");
     my $people_schema = $c->dbic_schema("CXGN::People::Schema");
 
@@ -89,7 +94,12 @@ sub delete :Path('/ajax/job/delete') Args(1) {
             sp_job_id => $sp_job_id
     });
 
+    if ($job->sp_person_id() ne $logged_user && $role ne "curator") {
+        die "You do not have permission to delete this job.\n";
+    }
+
     $job->delete();
+    $c->stash->{rest} = {success => 1};
 }
 
 sub cancel :Path('/ajax/job/cancel') Args(1) {
@@ -97,6 +107,9 @@ sub cancel :Path('/ajax/job/cancel') Args(1) {
     my $c = shift;
     my $sp_job_id = shift;
 
+    my $logged_user = $c->user() ? $c->user->get_object()->get_sp_person_id() : undef;
+    my $role = $c->user() ? $c->user->get_object()->get_user_type() : undef;
+
     my $bcs_schema = $c->dbic_schema("Bio::Chado::Schema");
     my $people_schema = $c->dbic_schema("CXGN::People::Schema");
 
@@ -106,5 +119,60 @@ sub cancel :Path('/ajax/job/cancel') Args(1) {
             sp_job_id => $sp_job_id
     });
 
+    if ($job->sp_person_id() ne $logged_user && $role ne "curator") {
+        die "You do not have permission to cancel this job.\n";
+    }
+
     $job->cancel();
+    $c->stash->{rest} = {success => 1};
+}
+
+sub delete_dead_jobs :Path('/ajax/job/delete_dead_jobs') Args(1) {
+    my $self = shift;
+    my $c = shift;
+    my $sp_person_id = shift;
+
+    my $bcs_schema = $c->dbic_schema("Bio::Chado::Schema");
+    my $people_schema = $c->dbic_schema("CXGN::People::Schema");
+
+    my $logged_user = $c->user() ? $c->user->get_object()->get_sp_person_id() : undef;
+    my $role = $c->user() ? $c->user->get_object()->get_user_type() : undef;
+    if ($sp_person_id ne $logged_user && $role ne "curator") {
+        die "You do not have permission to delete these job logs.\n";
+    }
+
+    CXGN::Job->delete_dead_jobs(
+        $bcs_schema,
+        $people_schema,
+        $sp_person_id
+    );
+    $c->stash->{rest} = {success => 1};
+}
+
+sub delete_older_than :Path('/ajax/job/delete_older_than') Args(2) {
+    my $self = shift;
+    my $c = shift;
+    my $older_than = shift;
+    my $sp_person_id = shift;
+
+    my $bcs_schema = $c->dbic_schema("Bio::Chado::Schema");
+    my $people_schema = $c->dbic_schema("CXGN::People::Schema");
+
+    my $logged_user = $c->user() ? $c->user->get_object()->get_sp_person_id() : undef;
+    my $role = $c->user() ? $c->user->get_object()->get_user_type() : undef;
+    if ($sp_person_id ne $logged_user && $role ne "curator") {
+        die "You do not have permission to delete these job logs.\n";
+    }
+
+    if ($older_than ne "one_week" &&  $older_than ne "one_month" && $older_than ne "six_months" && $older_than ne "one_year") {
+        die "Invalid time selection: $older_than.\n";
+    }
+
+    CXGN::Job->delete_jobs_older_than(
+        $bcs_schema,
+        $people_schema,
+        $sp_person_id,
+        $older_than
+    );
+    $c->stash->{rest} = {success => 1};
 }
